refactor file naming to match functions in file
