+++ src/lib.rs
@@ -1,4 +1,4 @@
-mod lines_to_params;
+mod lines_from_params;
 mod parser;
 mod utils;
 
@@ -9,10 +9,10 @@ use pyo3::types::PyDict;
 /// Main module for rust extensions with nari-act
 #[pymodule]
 fn nari_act_rust(py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_pyfunction!(lines_to_params::ability_from_params, m)?)?;
-    m.add_function(wrap_pyfunction!(lines_to_params::action_effect_from_params, m)?)?;
-    m.add_function(wrap_pyfunction!(lines_to_params::status_effect_from_params, m)?)?;
-    m.add_function(wrap_pyfunction!(lines_to_params::statuslist_from_params, m)?)?;
+    m.add_function(wrap_pyfunction!(lines_from_params::ability_from_params, m)?)?;
+    m.add_function(wrap_pyfunction!(lines_from_params::action_effect_from_params, m)?)?;
+    m.add_function(wrap_pyfunction!(lines_from_params::status_effect_from_params, m)?)?;
+    m.add_function(wrap_pyfunction!(lines_from_params::statuslist_from_params, m)?)?;
 
     let utils_mod = PyModule::new(py, "utils")?;
     utils_mod.add_function(wrap_pyfunction!(utils::date_from_cs_string, m)?)?;
